chore: changes to underwritertransferrequest
