+++ lib/zero.ts
@@ -19,5 +19,6 @@ export function createZeroKeeper(connection: ZeroConnection) {
 }
 
 export { getProvider, logger } from './deployment-utils';
+export * from './UnderwriterRequest'
 export * from './TransferRequest';
 export * from './MetaRequest';
